refactor: reorganize utility imports to use functions module and add convertToJson implementation
