+++ scripts/utils/functions.ts
