+++ scripts/utils/readAndWriteJson.ts
@@ -1,6 +1,6 @@
 import { readFile, writeFile } from 'fs/promises';
 
-import { convertToJson } from '../utils';
+import { convertToJson } from './functions';
 
 /**
  * Reads a file, converts its content to JSON, and writes the JSON to another file.
